gio: Add ListStore implementation
This patch adds a wrapper implementation for the GListStore class.

+++ gtk-bindings/src/nativeTest/kotlin/bindings.gtk/WindowTests.kt
@@ -16,6 +16,6 @@ class WindowTests : GtkTestBase() {
 
         window.present()
 
-        assertEquals(1, Window.toplevels.itemCount.toInt())
+        assertEquals(1, Window.toplevels.itemCount)
     }
 }
